Provide unit tests for PID binding feature
This commit creates fixtures and tests to ensure that there are unit
tests for the pid binding feature as it is currently understood in
Archivematica.

Also touched upon in this commit are changes to the parse_dataverse
tests. The tests were passing previously, but only because
additional data was not being added to the model we were using, but
as soon as we started to add new data, counts such as File counts
expected when we were not filtering correctly, were returning with
values other than those expected.